+++ app/Http/Controllers/EncomendasController.php
@@ -1,8 +1,9 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
+use Auth;
 use App\Models\Encomenda;
 use App\Models\Vendedor;
 use App\Models\Cliente;
@@ -29,14 +30,20 @@ class EncomendasController extends Controller
 
 
     public function create(){
-        $vendedor=Vendedor::all();
-        $cliente=Cliente::all();
-        $produto=Produto::all();
-        return view('encomendas.create',[
-            'vendedor'=>$vendedor,
-            'cliente'=>$cliente,
-            'produto'=>$produto
-        ]);
+        if(Gate::allows('admin')){
+            $vendedor=Vendedor::all();
+            $cliente=Cliente::all();
+            $produto=Produto::all();
+            return view('encomendas.create',[
+                'vendedor'=>$vendedor,
+                'cliente'=>$cliente,
+                'produto'=>$produto
+            ]);
+        }
+        else{
+            return redirect()->route('encomendas.index')
+            ->with('msg','Não têm permissão para aceder a area pretendida');
+        }
     }
 
 
@@ -48,31 +55,44 @@ class EncomendasController extends Controller
             'data'=>['nullable','date'],
             'observacoes'=>['nullable','min:3','max:255']
         ]);
-        $produto=$request->id_produto;
-        $encomenda=Encomenda::create($novaEncomenda);
-        $encomenda->produtos()->attach($produto);
-        return redirect()->route('encomendas.show',[
-            'id_encomenda'=>$encomenda->id_encomenda
-        ]);
+        if(Gate::allows('admin')){
+            $produto=$request->id_produto;
+            $encomenda=Encomenda::create($novaEncomenda);
+            $encomenda->produtos()->attach($produto);
+            return redirect()->route('encomendas.show',[
+                'id_encomenda'=>$encomenda->id_encomenda
+            ]);
+        }
+        else{
+            return redirect()->route('encomendas.index')
+            ->with('msg','Não têm permissão para aceder a area pretendida');
+        }
     }
 
     public function edit(Request $request){
         $id_encomenda = $request->id_encomenda;
         $encomenda=Encomenda::where('id_encomenda',$id_encomenda)->with(['produtos','clientes','vendedores'])->first();
-        $vendedor=Vendedor::all();
-        $cliente=Cliente::all();
-        $produto=Produto::all();
-        $produtosEncomendas=[];
-        foreach($encomenda->produtos as $produto){
-            $produtosEncomendas[]=$produto->id_produto;
+        if(Gate::allows('admin')){
+            $vendedor=Vendedor::all();
+            $cliente=Cliente::all();
+            $produtos=Produto::all();
+            $produtosEncomendas=[];
+            foreach($encomenda->produtos as $produto){
+                $produtosEncomendas[]=$produto->id_produto;
+            }
+            return view('encomendas.edit',[
+                'encomenda'=>$encomenda,
+                'vendedor'=>$vendedor,
+                'cliente'=>$cliente,
+                'produto'=>$produtos,
+                'produtosEncomendas'=>$produtosEncomendas
+            ]);
         }
-        return view('encomendas.edit',[
-            'encomenda'=>$encomenda,
-            'vendedor'=>$vendedor,
-            'cliente'=>$cliente,
-            'produto'=>$produto,
-            'produtosEncomendas'=>$produtosEncomendas
-        ]);
+        else{
+            return redirect()->route('encomendas.show')
+            ->with('msg','Não têm permissão para aceder a area pretendida');
+        }
+
     }
 
     public function update(Request $request){
@@ -85,34 +105,52 @@ class EncomendasController extends Controller
             'data'=>['nullable','date'],
             'observacoes'=>['nullable','min:3','max:255']
         ]);
-        $produto=$request->id_produto;
-        $editarEncomenda=$encomenda->update($editEncomenda);
-        $encomenda->produtos()->sync($produto);
-        return redirect()->route('encomendas.show',[
-            'id_encomenda'=>$encomenda->id_encomenda
-        ]);
+        if(Gate::allows('admin')){
+            $produto=$request->id_produto;
+            $editarEncomenda=$encomenda->update($editEncomenda);
+            $encomenda->produtos()->sync($produto);
+            return redirect()->route('encomendas.show',[
+                'id_encomenda'=>$encomenda->id_encomenda
+            ]);
+        }
+        else{
+            return redirect()->route('encomendas.show')
+            ->with('msg','Não têm permissão para aceder a area pretendida');
+        }
     }
 
     public function deleted(Request $request){
         $id_encomenda = $request->id_encomenda;
         $encomenda=Encomenda::where('id_encomenda',$id_encomenda)->with(['produtos','clientes','vendedores'])->first();
-        if(is_null($encomenda)){
-            return redirect()->route('encomendas.index')->with('msg','Não existe esta encomenda');
+        if(Gate::allows('admin')){
+            if(is_null($encomenda)){
+                return redirect()->route('encomendas.index')->with('msg','Não existe esta encomenda');
+            }
+            else{
+                return view('encomendas.delete',['encomenda'=>$encomenda]);
+            }
         }
         else{
-            return view('encomendas.delete',['encomenda'=>$encomenda]);
+            return redirect()->route('encomendas.show')
+            ->with('msg','Não têm permissão para aceder a area pretendida');
         }
     }
 
     public function destroy(Request $request){
         $id_encomenda = $request->id_encomenda;
         $encomenda=Encomenda::where('id_encomenda',$id_encomenda)->with(['produtos','clientes','vendedores'])->first();
-        if(is_null($encomenda)){
-            return redirect()->route('encomendas.index')->with('msg','A encomenda não existe');
+        if(Gate::allows('admin')){
+            if(is_null($encomenda)){
+                return redirect()->route('encomendas.index')->with('msg','A encomenda não existe');
+            }
+            else{
+                $encomenda->delete();
+                return redirect()->route('encomendas.index');
+            }
         }
         else{
-            $encomenda->delete();
-            return redirect()->route('encomendas.index');
+            return redirect()->route('encomendas.show')
+            ->with('msg','Não têm permissão para aceder a area pretendida');
         }
     }
 }
